box/v3: move manifest updating into the downloader itself
