+++ cmd/downloader/manifest/manifest.go
@@ -1,15 +1,13 @@
-package main
+package manifest
 
 import (
 	"bytes"
 	"io/ioutil"
 
 	"github.com/livepeer/catalyst/cmd/downloader/bucket"
-	"github.com/livepeer/catalyst/cmd/downloader/cli"
 	"github.com/livepeer/catalyst/cmd/downloader/constants"
 	"github.com/livepeer/catalyst/cmd/downloader/github"
 	"github.com/livepeer/catalyst/cmd/downloader/types"
-	"github.com/livepeer/catalyst/cmd/downloader/utils"
 	glog "github.com/magicsong/color-glog"
 	"gopkg.in/yaml.v3"
 )
@@ -28,22 +26,14 @@ func GenerateYamlManifest(manifest types.BoxManifest, path string) error {
 	return err
 }
 
-func Run(buildFlags types.BuildFlags) {
+// returns a manifest and boolean for whether we successfully wrote one
+func UpdateManifest(cliFlags types.CliFlags, m *types.BoxManifest) bool {
 	var projectInfo *types.ArtifactInfo
-
-	cliFlags, err := cli.GetCliFlags(buildFlags)
-	if err != nil {
-		glog.Fatal(err)
-	}
-	manifest, err := utils.ParseYamlManifest(cliFlags.ManifestFile, cliFlags.ManifestURL)
-	if err != nil {
-		glog.Fatal(err)
-	}
 
 	platform := cliFlags.Platform
 	architecture := cliFlags.Architecture
 
-	for _, service := range manifest.Box {
+	for _, service := range m.Box {
 		if service.Skip || service.SkipManifestUpdate {
 			continue
 		}
@@ -51,21 +41,19 @@ func Run(buildFlags types.BuildFlags) {
 			service.Strategy.Download = "github"
 		}
 		if service.Strategy.Download == "bucket" {
-			projectInfo = bucket.GetArtifactInfo(platform, architecture, manifest.Release, service)
+			projectInfo = bucket.GetArtifactInfo(platform, architecture, m.Release, service)
 		} else if service.Strategy.Download == "github" {
 			service.Release = constants.LatestTagReleaseName
-			projectInfo = github.GetArtifactInfo(platform, architecture, manifest.Release, service)
+			projectInfo = github.GetArtifactInfo(platform, architecture, m.Release, service)
 			service.Release = projectInfo.Version
 		}
 		glog.V(8).Infof("gh-version=%q, manifest-version=%q", projectInfo.Version, service.Release)
 	}
-	err = GenerateYamlManifest(*manifest, cliFlags.ManifestFile)
+	err := GenerateYamlManifest(*m, cliFlags.ManifestFile)
 
 	if err != nil {
-		glog.Fatal(err)
+		glog.Error(err)
+		return false
 	}
-}
-
-func main() {
-	Run(types.BuildFlags{Version: version})
+	return true
 }
