+++ cmd/downloader/types/types.go
@@ -36,6 +36,8 @@ type BuildFlags struct {
 type CliFlags struct {
 	SkipDownloaded bool
 	Cleanup        bool
+	UpdateManifest bool
+	Download       bool
 	DownloadPath   string
 	Platform       string
 	Architecture   string
